move the search form component into the character search scope
