+++ src/features/CharacterSearch/index.tsx
@@ -1,0 +1,1 @@
+export { default as SearchForm } from "./components/SearchForm";
